+++ test/unit/views/payment-method-view.js
@@ -104,8 +104,8 @@ describe('PaymentMethodView', function () {
       labelElement = this.context.element.querySelector('.braintree-method__label');
 
       expect(iconElement.getAttribute('xlink:href')).to.equal('#logoApplePay');
-      expect(labelElement.textContent).to.contain('Ending in ••92');
-      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('Apple Pay');
+      expect(labelElement.textContent).to.contain('Apple Pay');
+      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('');
       expect(iconContainer.classList.contains('braintree-method__logo@CLASSNAME')).to.be.false;
     });
 
@@ -125,8 +125,8 @@ describe('PaymentMethodView', function () {
 
       labelElement = this.context.element.querySelector('.braintree-method__label');
 
-      expect(labelElement.textContent).to.contain('Ending in ••92');
-      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('Apple Pay');
+      expect(labelElement.textContent).to.contain('Apple Pay');
+      expect(labelElement.querySelector('.braintree-method__label--small').textContent).to.equal('');
     });
   });